# web/openapi/natsume.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Natsume V2 Server API",
    "description": "Mounted Stage 5B operation IDs: getHealth, createSession, getSession, deleteSession, listSeats, listAccounts, listDevices, listBindings, revokeDevice, disableDevice, getCsvImport, createCsvImport, commitCsvImport, discardCsvImport, getProvisioningWindow, openProvisioningWindow, closeProvisioningWindow, createEnrollmentRequest.\nDeclared but not mounted in Stage 5B operation IDs: approveEnrollment, putCommand.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v2/enrollment-requests": {
      "post": {
        "operationId": "createEnrollmentRequest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Device credentials issued synchronously",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Server-generated canonical UUIDv7 correlation ID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentIssuedResponse"
                }
              }
            }
          },
          "202": {
            "description": "Enrollment request awaits approval or claim",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Server-generated canonical UUIDv7 correlation ID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentPendingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Enrollment request, CSR, SPKI, or protocol input",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Server-generated canonical UUIDv7 correlation ID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Provisioning window or Enrollment state conflict",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Server-generated canonical UUIDv7 correlation ID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request body exceeds the Enrollment ingress limit",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Server-generated canonical UUIDv7 correlation ID"
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "Server-generated canonical UUIDv7 correlation ID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CanonicalUuidV7": {
        "type": "string",
        "format": "uuid",
        "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
      },
...
      "EnrollmentHardwareIdentityQuality": {
        "type": "string",
        "enum": ["strong", "medium", "weak"]
      },
      "EnrollmentIssuedResponse": {
        "type": "object",
        "required": [
          "enrollment_request_id",
          "state",
          "device_id",
          "device_token",
          "gateway_leaf_der",
          "gateway_chain_der"
        ],
        "properties": {
          "device_id": {
            "$ref": "#/components/schemas/CanonicalUuidV7"
          },
          "device_token": {
            "type": "string",
            "maxLength": 43,
            "minLength": 43,
            "pattern": "^[A-Za-z0-9_-]{42}[AEIMQUYcgkosw048]$"
          },
          "enrollment_request_id": {
            "$ref": "#/components/schemas/CanonicalUuidV7"
          },
          "gateway_chain_der": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "byte"
            }
          },
          "gateway_leaf_der": {
            "type": "string",
            "format": "byte"
          },
          "state": {
            "$ref": "#/components/schemas/EnrollmentIssuedState"
          }
        },
        "additionalProperties": false
      },
      "EnrollmentIssuedState": {
        "type": "string",
        "enum": ["issued"]
      },
      "EnrollmentPendingResponse": {
        "type": "object",
        "required": ["enrollment_request_id", "state"],
        "properties": {
          "enrollment_request_id": {
            "$ref": "#/components/schemas/CanonicalUuidV7"
          },
          "state": {
            "$ref": "#/components/schemas/EnrollmentPendingState"
          }
        },
        "additionalProperties": false
      },
      "EnrollmentPendingState": {
        "type": "string",
        "enum": ["pending", "approved"]
      },
      "EnrollmentRequest": {
        "type": "object",
        "required": [
          "machine_hardware_id",
          "hardware_identity_quality",
          "gateway_csr_der",
          "gateway_spki_sha256",
          "client_version",
          "protocol_version"
        ],
        "properties": {
          "client_version": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[!-~]{1,64}$"
          },
          "gateway_csr_der": {
            "type": "string",
            "format": "byte",
            "maxLength": 43692,
            "minLength": 4,
            "pattern": "^(?:[A-Za-z0-9+/]{4})*(?:[A-Za-z0-9+/]{2}==|[A-Za-z0-9+/]{3}=)?$"
          },
          "gateway_spki_sha256": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "pattern": "^[0-9a-f]{64}$"
          },
          "hardware_identity_quality": {
            "$ref": "#/components/schemas/EnrollmentHardwareIdentityQuality"
          },
          "machine_hardware_id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-5[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
          },
          "protocol_version": {
            "type": "integer",
            "format": "int32",
            "maximum": 1,
            "minimum": 1
          }
        },
        "additionalProperties": false
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["title", "status", "code", "correlation_id"],
        "properties": {
          "code": {
            "type": "string"
          },
          "correlation_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Secure-natsume_session"
      }
    }
  }
}